requirements.txt
selenium
PyYAML
pytest
chromedrivermanager
pytest-ordering